backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
pydantic-settings
requests
beautifulsoup4
lxml
slowapi
python-jose
passlib[bcrypt]
openai